// src/enums/PenColors.ts
export enum PenColorMap {
  Black = 'Black',
  Blue = 'Blue',
  Yellow = 'Yellow',
  Red = 'Red',
  HighlightBlack = 'HighlightBlack',
  HighlightBlue = 'HighlightBlue',
  HighlightYellow = 'HighlightYellow',
  HighlightRed = 'HighlightRed',
}

export enum PenColors {
  Black = 'rgba(0,0,0,1)',
  Blue = 'rgba(0,0,255,1)',
  Yellow = 'rgba(255,255,0,1)',
  Red = 'rgba(255,0,0,1)',
  HighlightBlack = 'rgba(0,0,0,0.2)',
  HighlightBlue = 'rgba(0,0,255,0.3)',
  HighlightYellow = 'rgba(255,255,0,0.3)',
  HighlightRed = 'rgba(255,0,0,0.5)',
}

export enum PenColorLabels {
  Black = 'black',
  Blue = 'blue',
  Yellow = 'yellow',
  Red = 'red',
  HighlightBlack = 'highlight black',
  HighlightBlue = 'highlight blue',
  HighlightYellow = 'highlight yellow',
  HighlightRed = 'highlight red',
}